Restored round as it was

+++ openquake/hazardlib/gsim/gmpe_table.py
@@ -30,8 +30,7 @@ from openquake.baselib.python3compat import decode
 from openquake.hazardlib.const import TRT, StdDev
 from openquake.hazardlib import imt as imt_module
 from openquake.hazardlib.gsim.base import GMPE
-from numpy import round
-
+from openquake.baselib.python3compat import round
 
 _get_mean = CallableDict()
 
@@ -97,15 +96,15 @@ def _return_tables(self, mag, imt, which):
         iml_table = iml_table.reshape([n_d, n_m])
     else:
         if which == "IMLs":
-            periods = self.imls["T"][:].astype(np.float32)
+            periods = self.imls["T"][:]
             iml_table = self.imls["SA"][:]
         else:
-            periods = self.stddev["T"][:].astype(np.float32)
+            periods = self.stddev["T"][:]
             iml_table = self.stddev["SA"][:]
 
         low_period = round(periods[0], 7)
         high_period = round(periods[-1], 7)
-        period = round(np.float32(imt.period), 7)
+        period = round(imt.period, 7)
         if period < low_period or period > high_period:
             raise ValueError("Spectral period %.3f outside of valid range "
                              "(%.3f to %.3f)" % (imt.period, periods[0],
